pyproject.toml
[project]
name = "ai-storybook"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "annotated-types==0.7.0",
    "anyio==4.12.0",
    "asttokens==3.0.1",
    "blinker==1.9.0",
    "cachetools==6.2.4",
    "certifi==2025.11.12",
    "charset-normalizer==3.4.4",
    "click==8.3.1",
    "colorama==0.4.6",
    "decorator==5.2.1",
    "distro==1.9.0",
    "exceptiongroup==1.3.1",
    "executing==2.2.1",
    "filelock==3.20.1",
    "flask==3.1.2",
    "fsspec==2025.12.0",
    "google-auth==2.45.0",
    "google-genai==1.56.0",
    "h11==0.16.0",
    "hf-xet==1.2.0",
    "httpcore==1.0.9",
    "httpx==0.28.1",
    "huggingface-hub==1.2.3",
    "idna==3.11",
    "ipython==8.37.0",
    "itsdangerous==2.2.0",
    "jedi==0.19.2",
    "jinja2==3.1.6",
    "jiter==0.12.0",
    "markupsafe==3.0.3",
    "matplotlib-inline==0.2.1",
    "openai==2.14.0",
    "packaging==25.0",
    "parso==0.8.5",
    "pillow==12.0.0",
    "prompt-toolkit==3.0.52",
    "pure-eval==0.2.3",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pydantic==2.12.5",
    "pydantic-core==2.41.5",
    "pygments==2.19.2",
    "python-dotenv==1.2.1",
    "pyyaml==6.0.3",
    "requests==2.32.5",
    "rsa==4.9.1",
    "shellingham==1.5.4",
    "sniffio==1.3.1",
    "stack-data==0.6.3",
    "tenacity==9.1.2",
    "tqdm==4.67.1",
    "traitlets==5.14.3",
    "typer-slim==0.21.0",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.2",
    "urllib3==2.6.2",
    "wcwidth==0.2.14",
    "websockets==15.0.1",
    "werkzeug==3.1.4",
]
